math-speak/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "math-speak"
version = "0.1.0"
description = "Selected-text TTS for mathematical content (English + French)"
readme = "README.md"
requires-python = ">=3.11,<3.14"
license = { text = "MIT" }
authors = [{ name = "math-speak" }]
dependencies = [
    "pynput>=1.7.7",
    "pystray>=0.19",
    "Pillow>=10",
    "pylatexenc>=2.10",
    "piper-tts>=1.2.0",
    "sounddevice>=0.4.6",
    "httpx>=0.27",
    "tomli-w>=1.0",
]

[project.scripts]
math-speak = "math_speak.cli:main"
math-speakd = "math_speak.daemon:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
math_speak = ["data/*.json"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]
